fixes: chart selection and deselection
